Chore: 상수 지정, 위치이동, attempts 횟수 수정
- 상수 Names Max Length 지정
- StartRacingCar view → controller로 이동
- attemps 횟수 -1
