+++ src/js/model.js
@@ -1,7 +1,9 @@
+import {CAR} from "./constants/constants.js";
+
 export const validateCarNames = (namesArr) => {
   const namesError = namesArr
     .split(",")
-    .filter((name) => name.trim().length >= 6).length;
+    .filter((name) => name.trim().length >= CAR.NAME_MAX_LENGTH).length;
 
   if (!namesArr) return false;
   if (namesError > 0) return false;
